requirements.txt
python-docx
pyyaml
pytest
pdf2image
Pillow
